Refactor running for common use
